requirements.txt
openai==1.14.2
langchain==0.1.13
pydantic==2.6.4
langfuse==2.20.0
python-dotenv==1.0.1